+++ hyperactive/ViewModels/WTreeBranch.cs
@@ -55,7 +55,12 @@
                 index: 0)
             .ToArray();
 
-        private void Checkout() {
+        private async void Checkout() {
+            if (Parent.Status.WTreeStatus != WTreeStatus.Clean
+                && !await Dialog.Show(new Confirm("discard all changes and restore", subject: Name))) {
+                return;
+            }
+
             repo.Reset(ResetMode.Hard);
             Commands.Checkout(repo, LibGitBranch, new CheckoutOptions { CheckoutModifiers = CheckoutModifiers.Force });
             repo.RemoveUntrackedFiles();
